backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.2
sqlmodel==0.0.8
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai==1.36.0
email-validator==2.1.1
